Fix for correct usage in Blender 2.8
Along with bl_space_type changing from 'USER_PREFERENCES' to 'PREFERENCES' in Blender 2.8 the bl_context of 'data' no longer exists in the preferences dialog. I changed that to 'system' so this addon's controls would show up in the 'System' section of the preferences dialog.

There was no need to detect if this code was running in Blender 2.7x or 2.8x, as the 'blender' version metadata tuple at the top of the file limits it to 2.80 at a minimum. I removed the unnecessary version check to determin e which bl_space_type to use.